requirements.txt
# API and Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
streamlit>=1.29.0

# PDF Processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0
llama-index-core>=0.10.0
llama-parse>=0.1.0
llama-index-readers-file>=0.1.0

# AI and ML
openai>=1.3.0
langchain>=0.0.350
sentence-transformers>=2.2.2
pinecone>=0.2.0

# Google Cloud
google-cloud-bigquery>=3.13.0
google-cloud-storage>=2.13.0
google-cloud-firestore>=2.13.0

# Utilities
python-dotenv>=1.0.0
hashlib>=20081119
pandas>=2.1.4
numpy>=1.26.2
pytest>=7.4.3
black>=23.12.1
isort>=5.12.0
pydantic>=1.10.0
llama-index>=0.1.0
requests>=2.31.0
pydantic-settings==2.1.0 